OpenCL/Dijkstra.h
	OpenCL/device.h

+++ OpenCL/Dijkstra.h
@@ -6,4 +6,4 @@ void runDijkstraMultiGPUandCPU( cl_context gpuContext, cl_context cpuContext, Gr
                                 int *sourceVertices, int *outResultCosts, int numResults);
 
 void runDijkstraMultiGPU( cl_context gpuContext, GraphData* graph, int *sourceVertices,
-                          int *outResultCosts, int numResults );
+                          long *outResultCosts, int numResults );
